refactor run command to seperate module
